split zilean.ts into 4 files -> zilean,zileanUser,zileanComic,zileanStory
