+++ pages/comic/hub.tsx
@@ -3,7 +3,7 @@ import Head from "next/head";
 import Hub from "../../components/comic/hub";
 import Navbar from "../../components/navbar";
 import { AuthProvider } from "../../context/authcontext";
-import { getUserFromSession } from "../../util/zilean";
+import { getUserFromSession } from "../../util/zileanUser";
 
 interface Props {
     user: any;
