+++ pages/index.tsx
@@ -4,7 +4,7 @@ import PostLoginPage from "../components/postlogin";
 import LandingPage from "../components/landing";
 import Navbar from "../components/navbar";
 import { AuthProvider } from "../context/authcontext";
-import { getUserFromSession } from "../util/zilean";
+import { getUserFromSession } from "../util/zileanUser";
 
 interface Props {
     user: any;
