+++ pages/story/view/[id].tsx
@@ -3,7 +3,9 @@ import Head from "next/head";
 import ViewStory from "../../../components/story/view";
 import Navbar from "../../../components/navbar";
 import { AuthProvider } from "../../../context/authcontext";
-import { getUserFromSession, getStory, getUserFromID, getImage } from "../../../util/zilean";
+import { getImage } from "../../../util/zilean";
+import { getUserFromSession, getUserFromID } from "../../../util/zileanUser";
+import { getStory } from "../../../util/zileanStory";
 
 interface Props {
     user: any;
